Rewrite IntelliJ controller to use @previewjs/server
