+++ integrations/intellij/controller/package.json
@@ -20,13 +20,10 @@
   "devDependencies": {
     "@previewjs/core": "workspace:*",
     "@previewjs/loader": "workspace:*",
-    "@types/express": "4.17.13",
-    "@types/uuid": "8.3.4",
+    "@previewjs/server": "workspace:*",
     "esbuild": "0.14.49",
-    "express": "4.18.1",
     "rimraf": "3.0.2",
     "ts-node-dev": "2.0.0",
-    "typescript": "4.7.4",
-    "uuid": "8.3.2"
+    "typescript": "4.7.4"
   }
 }
